Create a ServerContext to store server level context.
