+++ extension/src/main/java/org/wildfly/testing/junit/extension/ModelControllerClientProducer.java
@@ -33,8 +33,8 @@ public class ModelControllerClientProducer implements ServerResourceProducer {
                     "Type %s is not assignable to %s".formatted(clazz.getName(), ModelControllerClient.class.getName()));
         }
 
-        // Get the server from WildFlyExtension and return as managed which will not allow shutting down the server
-        return WildFlyExtension.getServer(context)
+        // Get the server from ServerContext
+        return ServerContext.getServer(context)
                 .map(ServerManager::client)
                 .orElse(null);
     }
