+++ extension/src/main/java/org/wildfly/testing/junit/extension/WildFlyExtension.java
@@ -22,16 +22,13 @@ import org.junit.platform.commons.support.AnnotationSupport;
 import org.wildfly.plugin.tools.Deployment;
 import org.wildfly.plugin.tools.DeploymentResult;
 import org.wildfly.plugin.tools.UndeployDescription;
-import org.wildfly.plugin.tools.server.Configuration;
 import org.wildfly.plugin.tools.server.ServerManager;
 import org.wildfly.plugin.tools.server.ServerManagerListener;
 import org.wildfly.testing.junit.extension.annotation.ManualMode;
 import org.wildfly.testing.junit.extension.annotation.ServerGroup;
 import org.wildfly.testing.junit.extension.annotation.WildFlyDomainTest;
 import org.wildfly.testing.junit.extension.annotation.WildFlyTest;
-import org.wildfly.testing.junit.extension.api.DomainConfigurationFactory;
 import org.wildfly.testing.junit.extension.api.ServerConfiguration;
-import org.wildfly.testing.junit.extension.api.StandaloneConfigurationFactory;
 
 /**
  * JUnit extension that manages WildFly server lifecycle and per-test-class deployments.
@@ -48,10 +45,9 @@ import org.wildfly.testing.junit.extension.api.StandaloneConfigurationFactory;
 public class WildFlyExtension implements BeforeAllCallback, AfterAllCallback {
     private static final Logger LOGGER = Logger.getLogger(WildFlyExtension.class);
 
-    private static final String SERVER_KEY = "wildfly.server";
     private static final String SERVER_LISTENER_KEY = "wildfly.server.listener";
-    private static final ExtensionContext.Namespace SERVER_NAMESPACE = ExtensionContext.Namespace
-            .create("WildFly.Server");
+    private static final ExtensionContext.Namespace CLASS_STORE_NAMESPACE = ExtensionContext.Namespace
+            .create("WildFly.ClassStore");
 
     @Override
     public void beforeAll(final ExtensionContext context) throws Exception {
@@ -59,7 +55,7 @@ public class WildFlyExtension implements BeforeAllCallback, AfterAllCallback {
         validateTestModeAnnotations(context);
 
         // Start server (if not already started) - shared across all test classes
-        final ServerManager serverManager = getOrCreateServerManager(context);
+        final ServerManager serverManager = ServerContext.getOrCreateServerManager(context);
 
         final Optional<ManualMode> manualMode = TestSupport.getManualMode(context);
 
@@ -86,7 +82,7 @@ public class WildFlyExtension implements BeforeAllCallback, AfterAllCallback {
                 if (serverManager.isRunning()) {
                     LOGGER.debugf("Shutting down server for manual mode test %s", context.getRequiredTestClass()
                             .getName());
-                    stopServer(context, serverManager);
+                    ServerContext.stopServer(context, serverManager);
                 }
                 final var listener = new ExtensionServerManagerListener(context, serverManager);
                 getClassStore(context).put(SERVER_LISTENER_KEY, listener);
@@ -97,7 +93,7 @@ public class WildFlyExtension implements BeforeAllCallback, AfterAllCallback {
 
     @Override
     public void afterAll(final ExtensionContext context) {
-        final Optional<ServerManager> opt = getServer(context);
+        final Optional<ServerManager> opt = ServerContext.getServer(context);
         if (opt.isEmpty()) {
             return;
         }
@@ -108,40 +104,6 @@ public class WildFlyExtension implements BeforeAllCallback, AfterAllCallback {
             return;
         }
         undeploy(context, serverManager, deploymentInfo.get());
-    }
-
-    /**
-     * Gets or starts the shared WildFly server instance.
-     * The server is stored in the root context store and shared across all test classes.
-     *
-     * @param context the extension context
-     *
-     * @return the server manager
-     */
-    @SuppressWarnings("resource")
-    private ServerManager getOrCreateServerManager(final ExtensionContext context) {
-        final ExtensionContext.Store store = getGlobalStore(context);
-
-        return store.computeIfAbsent(SERVER_KEY, key -> {
-            // Start the server
-            final ServerManager serverManager = createServer(context);
-            // Return a CloseableResource that stops server on cleanup
-            return new ServerResource(serverManager, context);
-        }, ServerResource.class).get();
-    }
-
-    /**
-     * Gets the already-started server, if available.
-     * Package-private to allow access from {@link ServerManagerProducer}.
-     *
-     * @param context the extension context
-     *
-     * @return the server manager, or empty if not started
-     */
-    static Optional<ServerManager> getServer(final ExtensionContext context) {
-        final ExtensionContext.Store store = getGlobalStore(context);
-        final ServerResource resource = store.get(SERVER_KEY, ServerResource.class);
-        return Optional.ofNullable(resource).map(ServerResource::get);
     }
 
     /**
@@ -162,41 +124,6 @@ public class WildFlyExtension implements BeforeAllCallback, AfterAllCallback {
             throw new JUnitException(
                     "Test class %s cannot have both @WildFlyTest and @WildFlyDomainTest. Use only one to specify the test mode."
                             .formatted(testClass.getName()));
-        }
-    }
-
-    private ServerManager createServer(final ExtensionContext context) {
-        // Determine configuration based on launch type
-        final Class<?> testClass = context.getRequiredTestClass();
-        final boolean hasDomainTest = AnnotationSupport
-                .findAnnotation(testClass, WildFlyDomainTest.class).isPresent();
-
-        final Configuration<?> configuration;
-        if (hasDomainTest) {
-            configuration = DomainConfigurationFactory.create()
-                    .configuration(context);
-        } else {
-            configuration = StandaloneConfigurationFactory.create()
-                    .configuration(context);
-        }
-        return ServerManager.of(configuration);
-    }
-
-    /**
-     * Stops the WildFly server gracefully.
-     *
-     * @param context       the extension context
-     * @param serverManager the server manager to stop
-     */
-    private void stopServer(final ExtensionContext context, final ServerManager serverManager) {
-        try {
-            // Get timeout from configuration (defaults to 60 seconds)
-            final long timeout = ServerConfiguration.timeout(context);
-            serverManager.shutdown(timeout);
-        } catch (IOException e) {
-            // If graceful shutdown fails, kill the server
-            serverManager.kill();
-            LOGGER.error("Failed to stop server", e);
         }
     }
 
@@ -282,10 +209,6 @@ public class WildFlyExtension implements BeforeAllCallback, AfterAllCallback {
         DeploymentContext.remove(context);
     }
 
-    private static ExtensionContext.Store getGlobalStore(final ExtensionContext context) {
-        return context.getRoot().getStore(ExtensionContext.StoreScope.LAUNCHER_SESSION, SERVER_NAMESPACE);
-    }
-
     private static ExtensionContext.Store getClassStore(final ExtensionContext context) {
         // We want to store on the class context, attempt to determine which that context
         final ExtensionContext usingContext;
@@ -294,7 +217,7 @@ public class WildFlyExtension implements BeforeAllCallback, AfterAllCallback {
         } else {
             usingContext = context;
         }
-        return usingContext.getStore(SERVER_NAMESPACE);
+        return usingContext.getStore(CLASS_STORE_NAMESPACE);
     }
 
     private static Optional<Archive<?>> resolveDeployment(final ExtensionContext context) {
@@ -331,28 +254,6 @@ public class WildFlyExtension implements BeforeAllCallback, AfterAllCallback {
 
         return serverGroupAnnotation.map(serverGroup -> Set.of(serverGroup.value())).orElseGet(Set::of);
 
-    }
-
-    /**
-     * Wrapper for ServerManager that implements AutoCloseable for automatic cleanup.
-     */
-    private class ServerResource implements AutoCloseable {
-        private final ServerManager serverManager;
-        private final ExtensionContext context;
-
-        ServerResource(final ServerManager serverManager, final ExtensionContext context) {
-            this.serverManager = serverManager;
-            this.context = context;
-        }
-
-        ServerManager get() {
-            return serverManager;
-        }
-
-        @Override
-        public void close() {
-            stopServer(context, serverManager);
-        }
     }
 
     private class ExtensionServerManagerListener implements ServerManagerListener, AutoCloseable {
